shijue_exe_4/requirements.txt
# 基础依赖
torch>=1.7.0
torchvision>=0.8.1

# 图像处理
opencv-python>=4.5.1
numpy>=1.19.2

# YOLOv5依赖
ultralytics>=8.0.0

# 可视化
matplotlib>=3.3.4
seaborn>=0.11.1

# 数据处理
tqdm>=4.60.0
PyYAML>=5.4.1